Adds a stackable component as well as some signal file management (#7853)
… things
## About The Pull Request

Title T

## Why It's Good For The Game

Likely gonna refactor existing stacks into using it later

## Changelog
:cl:
/:cl:

---------

+++ code/modules/explorer_drone/exodrone.dm
@@ -462,6 +462,9 @@ GLOBAL_LIST_EMPTY(exodrone_launchers)
 	var/uses = 5
 
 /obj/item/fuel_pellet/use(used)
+	. = ..()
+	if(!.)
+		return
 	uses--
 	if(uses <= 0)
 		qdel(src)
